Privacy_Handler_for_LLMs/Dependent_Model/requirements.txt
# Deep Learning Framework
tensorflow>=2.18.0
# For GPU support (optional):
# tensorflow-gpu>=2.18.0

# Data Processing
numpy==1.26.4
pandas>=2.1.0
scikit-learn==1.3.0

# Text Processing
nltk==3.8.1
regex==2023.6.3

# Synthetic Data Generation
faker==19.3.0

# API Integration
requests==2.31.0
google-generativeai==0.7.2
python-dotenv==1.0.0

# Visualization and Analysis
matplotlib==3.7.2
seaborn==0.12.2

# Progress Tracking
tqdm==4.65.0

# Configuration Management
pyyaml==6.0.1

# Testing
pytest==7.4.0

# Logging
loguru==0.7.0

# Performance Monitoring
memory-profiler==0.61.0
psutil==5.9.5

# Additional utilities
jsonschema==4.19.0

# Web Interface
streamlit>=1.28.0